shifts_website/requirements.txt
Flask>=2.0
Authlib>=1.2
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
Flask-APScheduler==1.13.1
APScheduler==3.11.0
